remove unnecessary dashboard section files and simplify page composition
