--- components/dashboard/sections/index.ts
@@ -1,5 +1,0 @@
-export * from './IdentityStatusSection';
-export * from './NextActionsSection';
-export * from './AnalyticsSection';
-export * from './MotivationSection';
-export * from './UtilitiesSection';
