+++ pages/dashboard/index.tsx
@@ -1,7 +1,10 @@
 import type { NextPage } from 'next';
+import Link from 'next/link';
 import dynamic from 'next/dynamic';
 import { useMemo, useState } from 'react';
 import useSWR from 'swr';
+import { Card } from '@/components/design-system/Card';
+import { ProgressBar } from '@/components/design-system/ProgressBar';
 import useDashboard, {
   useEstimatedBandScore,
   useImprovementGraph,
@@ -9,18 +12,18 @@ import useDashboard, {
   useStrengthsWeaknesses,
   useStudyStreak,
 } from '@/hooks/useDashboard';
-import {
-  AnalyticsSection,
-  IdentityStatusSection,
-  MotivationSection,
-  NextActionsSection,
-  UtilitiesSection,
-} from '@/components/dashboard/sections';
 
 const AICommandCenter = dynamic(
   () => import('@/components/dashboard/AICommandCenter').then((mod) => mod.AICommandCenter),
   {
     loading: () => <p className="text-xs text-muted-foreground">Loading AI command center…</p>,
+  },
+);
+
+const TrendChart = dynamic(
+  () => import('@/components/charts/TrendChart').then((mod) => mod.TrendChart),
+  {
+    loading: () => <p className="text-xs text-muted-foreground">Loading chart…</p>,
   },
 );
 
@@ -57,30 +60,185 @@ const DashboardPage: NextPage = () => {
 
   return (
     <>
-      <section className="space-y-6">
+      <section className="space-y-12">
         {error ? (
           <div className="rounded-xl border border-red-200 bg-red-50 p-6 text-sm text-red-700 dark:border-red-900/40 dark:bg-red-900/20 dark:text-red-200">
             Failed to load dashboard analytics.
           </div>
         ) : null}
 
-        <IdentityStatusSection
-          prediction={prediction}
-          score={score}
-          streak={streak}
-          heatmap={heatmap}
-          strengths={strengths}
-          weaknesses={weaknesses}
-        />
+        <section className="space-y-6">
+          <header className="space-y-2">
+            <h1 className="text-2xl font-semibold">AI Learning Dashboard</h1>
+            <p className="text-sm text-muted-foreground">
+              Your enterprise learning profile with score trend, skill heatmap, and study momentum.
+            </p>
+          </header>
 
-        <NextActionsSection recommendations={recommendations?.nextExercises ?? []} />
-        <AnalyticsSection points={points} />
-        <MotivationSection streak={streak} />
-        <UtilitiesSection
-          whatIfWriting={whatIfWriting}
-          setWhatIfWriting={setWhatIfWriting}
-          whatIf={whatIf}
-        />
+          <div className="grid gap-6 md:grid-cols-3">
+            <Card className="p-6" interactive>
+              <p className="text-xs uppercase tracking-wide text-muted-foreground">
+                Predicted Band Score
+              </p>
+              <p className="mt-2 text-3xl font-semibold">
+                {typeof prediction?.predictedBand === 'number'
+                  ? prediction.predictedBand.toFixed(1)
+                  : typeof score === 'number'
+                    ? score.toFixed(1)
+                    : '—'}
+              </p>
+              <p className="mt-1 text-xs text-muted-foreground">
+                Confidence ±
+                {typeof prediction?.confidenceInterval === 'number'
+                  ? prediction.confidenceInterval
+                  : 0.5}{' '}
+                · Trend: {prediction?.trend ?? 'stable'}
+              </p>
+            </Card>
+
+            <Card className="p-6" interactive>
+              <p className="text-xs uppercase tracking-wide text-muted-foreground">Study Streak</p>
+              <p className="mt-2 text-3xl font-semibold">🔥 {streak} days</p>
+              <p className="mt-1 text-xs text-muted-foreground">
+                Consistency is the fastest path to target band gains.
+              </p>
+            </Card>
+
+            <Card className="p-6" interactive>
+              <p className="text-xs uppercase tracking-wide text-muted-foreground">
+                Billing & Usage
+              </p>
+              <p className="mt-2 text-sm text-muted-foreground">
+                Track plan usage meters and invoices in one place.
+              </p>
+              <Link
+                href="/settings/billing"
+                className="mt-3 inline-flex text-sm font-medium text-primary hover:underline"
+              >
+                View details
+              </Link>
+            </Card>
+          </div>
+
+          <div className="grid gap-6 lg:grid-cols-2">
+            <Card className="p-6" interactive>
+              <h2 className="text-base font-semibold">Skill Heatmap</h2>
+              <div className="mt-3 space-y-2">
+                {heatmap.map((item) => (
+                  <div key={item.skill} className="space-y-1">
+                    <div className="flex justify-between text-xs">
+                      <span>{item.skill}</span>
+                      <span>{item.score.toFixed(1)}/10</span>
+                    </div>
+                    <ProgressBar value={Math.max(0, Math.min(100, item.score * 10))} />
+                  </div>
+                ))}
+              </div>
+            </Card>
+
+            <Card className="p-6" interactive>
+              <h2 className="text-base font-semibold">Strengths & Weaknesses</h2>
+              <div className="mt-3 grid grid-cols-2 gap-6 text-sm">
+                <div>
+                  <p className="font-medium text-emerald-600">Strengths</p>
+                  <ul className="mt-1 space-y-1 text-muted-foreground">
+                    {strengths.map((s) => (
+                      <li key={s.skill}>
+                        {s.skill} ({s.score.toFixed(1)})
+                      </li>
+                    ))}
+                  </ul>
+                </div>
+                <div>
+                  <p className="font-medium text-amber-600">Weaknesses</p>
+                  <ul className="mt-1 space-y-1 text-muted-foreground">
+                    {weaknesses.map((s) => (
+                      <li key={s.skill}>
+                        {s.skill} ({s.score.toFixed(1)})
+                      </li>
+                    ))}
+                  </ul>
+                </div>
+              </div>
+            </Card>
+          </div>
+        </section>
+
+        <section className="grid gap-6 lg:grid-cols-3">
+          <Card className="p-6 lg:col-span-2" interactive>
+            <h2 className="text-base font-semibold">Recommended for You</h2>
+            <p className="mt-1 text-xs text-muted-foreground">
+              Today&apos;s recommendations based on your latest skill profile.
+            </p>
+            <div className="mt-3 space-y-2">
+              {(recommendations?.nextExercises ?? []).slice(0, 5).map((item: any) => (
+                <div key={item.taskId} className="rounded-lg border border-border/60 p-3 text-sm">
+                  <p className="font-medium">{item.title}</p>
+                  <p className="text-xs text-muted-foreground">
+                    {item.module} · {item.type} · {item.reason}
+                  </p>
+                </div>
+              ))}
+              {!(recommendations?.nextExercises ?? []).length ? (
+                <p className="text-xs text-muted-foreground">No personalized exercises yet.</p>
+              ) : null}
+            </div>
+          </Card>
+        </section>
+
+        <section>
+          <Card className="p-6" interactive>
+            <TrendChart
+              title="Improvement Graph"
+              points={(points.length ? points : [{ label: 'No data', band: 0 }]).map((p) => ({
+                label: p.label,
+                value: p.band,
+              }))}
+            />
+          </Card>
+        </section>
+
+        <section>
+          <Card className="p-6" interactive>
+            <h2 className="text-base font-semibold">Motivation & Momentum</h2>
+            <p className="mt-1 text-sm text-muted-foreground">
+              {streak < 3
+                ? 'Your streak is low—complete one quick task today to rebuild consistency.'
+                : 'Great consistency—keep your streak alive to compound your score growth.'}
+            </p>
+          </Card>
+        </section>
+
+        <section>
+          <Card className="p-6" interactive>
+            <h2 className="text-base font-semibold">What-if simulator</h2>
+            <p className="mt-1 text-xs text-muted-foreground">If Writing improves to:</p>
+            <input
+              type="range"
+              min={4}
+              max={9}
+              step={0.5}
+              value={whatIfWriting}
+              onChange={(event) => setWhatIfWriting(Number(event.target.value))}
+              className="mt-3 w-full"
+            />
+            <p className="mt-2 text-sm">
+              Writing target: <span className="font-semibold">{whatIfWriting.toFixed(1)}</span>
+            </p>
+            <p className="mt-1 text-sm">
+              Predicted:{' '}
+              <span className="font-semibold">
+                {typeof whatIf?.predictedBand === 'number' ? whatIf.predictedBand.toFixed(1) : '—'}
+              </span>
+            </p>
+            <p className="mt-1 text-xs text-emerald-600">
+              Potential uplift:{' '}
+              {typeof whatIf?.uplift === 'number'
+                ? `${whatIf.uplift >= 0 ? '+' : ''}${whatIf.uplift.toFixed(2)}`
+                : '—'}
+            </p>
+          </Card>
+        </section>
 
         {isLoading ? (
           <p className="text-xs text-muted-foreground">Loading dashboard analytics…</p>
